collector/requirements.txt
paho-mqtt<2.0.0
fastapi
uvicorn
requests
websockets
